requirements.txt
fastapi
uvicorn
shapely
pyproj
requests
typing
google-genai
json
os
pydantic